another small bugfix, same relation

+++ nn.py
@@ -47,10 +47,10 @@ class NeuralNetwork:
         p = y_pred
         pi = y_true
 
-        zero = tf.zeros(shape = tf.shape(pi))
+        zero = tf.zeros(shape = tf.shape(pi), dtype = np.float32)
         where = tf.equal(pi, zero)
 
-        negatives = tf.fill(tf.shape(pi), -100) 
+        negatives = tf.fill(tf.shape(pi), -100.0) 
         p = tf.where(where, negatives, p)
 
         loss = tf.nn.softmax_cross_entropy_with_logits(labels = pi, logits = p)
